+++ projects/plugins/super-cache/wp-cache-phase2.php
@@ -2570,9 +2570,9 @@ function prune_super_cache( $directory, $force = false, $rename = false ) {
 				prune_super_cache( $entry, $force, $rename );
 				// If entry is a directory, AND it's not a protected one, AND we're either forcing the delete, OR the file is out of date,
 				if ( is_dir( $entry ) && ! in_array( $entry, $protected_directories ) && ( $force || @filemtime( $entry ) + $cache_max_time <= $now ) ) {
+					$donotdelete = false;
 					// if the directory isn't empty can't delete it
 					if ( $handle = @opendir( $entry ) ) {
-						$donotdelete = false;
 						while ( ! $donotdelete && ( $file = @readdir( $handle ) ) !== false ) {
 							if ( $file == '.' || $file == '..' ) {
 								continue;
@@ -3096,6 +3096,11 @@ function wp_cache_post_edit( $post_id ) {
 		return $post_id;
 	}
 
+	// Allow plugins to reject cache clears for specific posts.
+	if ( ! apply_filters( 'wp_super_cache_clear_post_cache', true, $post ) ) {
+		return $post_id;
+	}
+
 	// Some users are inexplicibly seeing this error on scheduled posts.
 	// define this constant to disable the post status check.
 	if ( ! defined( 'WPSCFORCEUPDATE' ) && ! in_array( get_post_status( $post ), array( 'publish', 'private' ), true ) ) {
@@ -3109,6 +3114,8 @@ function wp_cache_post_edit( $post_id ) {
 		wp_cache_debug( "wp_cache_post_edit: Clearing cache $blog_cache_dir and {$cache_path}supercache/ on post edit per config.", 2 );
 		prune_super_cache( $blog_cache_dir, true );
 		prune_super_cache( get_supercache_dir(), true );
+
+		do_action( 'wp_cache_cleared' );
 	} else {
 		$action = current_filter();
 		wp_cache_debug( "wp_cache_post_edit: Clearing cache for post $post_id on {$action}", 2 );