+++ projects/plugins/super-cache/CHANGELOG.md
@@ -5,6 +5,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.3-beta] - 2023-01-23
+### Added
+- Added new filters to set mod_expires rules and HTTP headers in the cache htaccess file. [#28031]
+
+### Fixed
+- Fixed an issue that caused wp-config.php file permissions to change. [#28164]
+- Fixed missing missing action 'wp_cache_cleared' when clearing the cache on post update. [#28481]
+
 ## [1.9.2-beta] - 2022-12-09
 ### Added
 - Added a dismissable option to install Jetpack Boost. [#26702]
@@ -627,6 +635,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Misc fixes
 
+[1.9.3-beta]: https://github.com/Automattic/wp-super-cache/compare/v1.9.2-beta...v1.9.3-beta
 [1.9.2-beta]: https://github.com/Automattic/wp-super-cache/compare/v1.9.1...v1.9.2-beta
 [1.9.1]: https://github.com/Automattic/wp-super-cache/compare/v1.9.0...v1.9.1
 [1.9.0]: https://github.com/Automattic/wp-super-cache/compare/v1.8...v1.9